+++ functions/Close-MySQLiteDB.ps1
@@ -15,6 +15,8 @@ Function Close-MySQLiteDB {
     )
     Begin {
         Write-Verbose "[$((Get-Date).TimeOfDay)] $($MyInvocation.MyCommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Running under PowerShell version $($PSVersionTable.PSVersion)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Detected culture $(Get-Culture)"
     } #begin
 
     Process {