+++ functions/New-MySQLiteDB.ps1
@@ -21,7 +21,8 @@ Function New-MySQLiteDB {
 
     Begin {
         Write-Verbose "[$((Get-Date).TimeOfDay)] $($MyInvocation.MyCommand)"
-
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Running under PowerShell version $($PSVersionTable.PSVersion)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Detected culture $(Get-Culture)"
         $db = resolvedb $Path
 
         If (($db.exists) -AND (-not $Force)) {