+++ changelog.md
@@ -2,6 +2,20 @@
 
 ## [Unreleased]
 
+## [0.14.0] - 2024-08-22
+
+### Changed
+
+- Updated `Import-MySqliteDB` to use an existing database. The new parameter is called `UseExisting` with an alias of `Append`. _This may be a breaking change_.
+- Updated functions with additional verbose output to capture PowerShell version and culture.
+- Help updates
+- Updated `README.md`
+
+### Fixed
+
+- Modified helper functions to store DateTime values as `yyyy-MM-dd HH:mm:ss` which should better handle culture-related problems. [Issue #23](https://github.com/jdhitsolutions/mySQLite/issues/23) _This may be a breaking change_.
+
+
 ## [0.13.0] - 2024-02-22
 
 ### Changed
@@ -159,7 +173,8 @@ This release is identical in files to v0.11.0 published on the PowerShell Galler
 - Added `Get-MySQLiteTable`
 - restructured module layout
 
-[Unreleased]: https://github.com/jdhitsolutions/MySQLite/compare/v0.13.0..HEAD
+[Unreleased]: https://github.com/jdhitsolutions/MySQLite/compare/v0.14.0..HEAD
+[0.14.0]: https://github.com/jdhitsolutions/MySQLite/compare/v0.13.0..v0.14.0
 [0.13.0]: https://github.com/jdhitsolutions/MySQLite/compare/v0.12.0..v0.13.0
 [v0.12.0]: https://github.com/jdhitsolutions/MySQLite/compare/v0.11.2..v0.12.0
 [v0.11.2]: https://github.com/jdhitsolutions/MySQLite/compare/v0.11.1..v0.11.2